requirements.txt
agora_token_builder==1.0.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
captcha==0.5.0
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==41.0.5
daphne==4.0.0
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.58.1
django-appconf==1.0.5
django-compressor==4.4
django-environ==0.11.2
django-extensions==3.2.3
django-recaptcha==3.0.0
django-redis==5.4.0
django-sass-processor==1.3
environ==1.0
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jwt==1.3.1
libsass==0.22.0
MarkupSafe==2.1.3
msgpack==1.0.7
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyforkurento==0.1.7
PyJWT==2.8.0
pyOpenSSL==23.3.0
python3-openid==3.2.0
rcssmin==1.1.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rjsmin==1.2.1
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.4
urllib3==2.0.7
websocket-client==0.57.0
Werkzeug==3.0.1
zope.interface==6.1
